Refactor import path for createAuthProfileExp to new utility file in PacAuthUtil
